MEDOMIT/codebase/requirements.txt
nltk~=3.8.1
tqdm~=4.65.0
pandas~=1.5.1
jinja2~=3.1.2
google~=3.0.0
configargparse~=1.4
openai~=0.27.7
tiktoken~=0.3.3
numpy~=1.23.4
langchain[llms]~=0.0.317
pydantic~=1.10.11
xlsxwriter~=3.1.1
pyyaml~=6.0.1
